Accept a list of local environment directories that wont be touched by the agent (for local development on test masters)

+++ util/puppetenvsh/puppetenvsh.rb
@@ -20,6 +20,8 @@ module MCollective
                 @upstream = config.pluginconf.fetch('puppetenvsh.upstream', 'origin')
                 # Username for connecting to the remote
                 @username = config.pluginconf.fetch('puppetenvsh.username', 'git')
+                # Ignore local environments
+                @local_environments = config.pluginconf.fetch('puppetenvsh.local_environments', 'live')
                 # Whether to use librarian-puppet to manage modules
                 @use_librarian = config.pluginconf.fetch('puppetenvsh.use_librarian', "false")
                 # Paths for binaries used by this agent
@@ -35,6 +37,9 @@ module MCollective
                 # Convert string booleans to real booleans
                 @use_librarian = !! (@use_librarian =~ /^1|true|yes/)
                 @use_ruby193   = !! (@use_ruby193   =~ /^1|true|yes/)
+
+                # Convert lists
+                @local_environments = @local_environments.split(/,/)
 
                 @ruby_env = ""
                 if @use_ruby193
@@ -65,6 +70,8 @@ module MCollective
             # * submodules will be updated if present
             # * librarian will be run if enabled
             def add(name)
+                return false, "#{name} is a protected local environment" if @local_environments.include?(name)
+
                 workdir = File.join(@basedir, name)
                 command = "#{@new_workdir.shellescape} #{@master_repo_path.shellescape} #{workdir.shellescape} #{name.shellescape} >/dev/null 2>&1"
                 output = ""
@@ -90,6 +97,8 @@ module MCollective
             # * Submodules will be updated if present
             # * librarian will be run if enabeld
             def update(name)
+                return false, "#{name} is a protected local environment" if @local_environments.include?(name)
+
                 workdir = File.join(@basedir, name)
                 command = "#{@git.shellescape} reset --hard #{@upstream.shellescape}/#{name.shellescape} 2>&1 >/dev/null"
                 output = ""
@@ -115,8 +124,12 @@ module MCollective
             #
             # * It's assumed existence of the environment has already been checked
             def rm(name)
+                return false, "#{name} is a protected local environment" if @local_environments.include?(name)
+
                 workdir = File.join(@basedir, name)
                 FileUtils.remove_entry_secure(workdir, true)
+
+                return true
             end
 
             # Updates all environemnts
@@ -137,6 +150,8 @@ module MCollective
                 messages = ""
 
                 upstream_branches.each do |branch|
+                    next if @local_environments.include?(branch)
+
                     environment = File.join(@basedir, branch)
                     if File.directory?(environment)
                         result, output = update(branch)
@@ -168,6 +183,8 @@ module MCollective
 
                 branches = upstream_branches
                 list.each do |environment|
+                    next if @local_environments.include?(environment)
+
                     if not branches.include?(environment)
                         rm(environment)
                         results[:removed] << environment
@@ -232,12 +249,14 @@ module MCollective
             #
             # * It must only consist of alpahnumeric characters and underscores
             # * It must refer to an existing branch in the upstream remote
-            def validate_environment_name(name)
+            def validate_environment_name(name, require_branch=true)
                 # Verify the name is valid as a puppet environment
                 return false unless name.match(/^[a-zA-Z0-9_]+$/)
 
                 # Verify the name corresponds with an existing git branch
-                return false unless upstream_branches.include? name
+                if require_branch
+                    return false unless upstream_branches.include? name
+                end
 
                 return true
             end
